requirements.txt
Django>=4.1.3,<4.2
djangorestframework>=3.14.0,<3.15
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.19.<0.2
